requirements.txt
python-decouple~=3.8
flask~=2.3.3
flask-cors~=4.0.0